move simple test form creation to util
