+++ corehq/form_processor/utils/__init__.py
@@ -7,6 +7,9 @@ from .xform import (
     extract_meta_user_id,
     convert_xform_to_json,
     adjust_datetimes,
+    get_simple_form_xml,
+    get_simple_wrapped_form,
+    TestFormMetadata,
 )
 
 from .metadata import (
