+++ testapps/test_elasticsearch/tests/test_xform_es.py
@@ -5,7 +5,7 @@ from django.test import SimpleTestCase
 from elasticsearch.exceptions import ConnectionError
 
 from corehq.apps.es import FormES
-from corehq.form_processor.tests.utils import TestFormMetadata
+from corehq.form_processor.utils import TestFormMetadata
 from corehq.pillows.xform import XFormPillow
 from corehq.util.test_utils import make_es_ready_form, trap_extra_setup
 from pillowtop.tests import require_explicit_elasticsearch_testing
